pyproject.toml
[project]
name = "speckit-docs"
version = "0.1.0"
description = "AI-driven documentation generation system for spec-kit projects"
readme = "README.md"
authors = [
    { name = "driller", email = "eleshis@gmail.com" }
]
requires-python = ">=3.11"
keywords = ["spec-kit", "documentation", "sphinx", "mkdocs", "markdown"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Documentation",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "sphinx>=7.0",
    "myst-parser>=2.0",
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "markdown-it-py>=3.0",
    "GitPython>=3.1",
    "Jinja2>=3.1",
    "ruamel.yaml>=0.18",
    "specify-cli",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "black>=24.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
    "pyfakefs>=5.0",
]

[project.scripts]
speckit-docs = "speckit_docs.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-ra -q --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "contract: Contract tests",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
fixable = ["ALL"]
ignore = ["E501"]  # Ignore line length errors for Japanese messages

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
implicit_reexport = false
strict_equality = true

[tool.uv.sources]
specify-cli = { git = "https://github.com/github/spec-kit.git" }
